Move process modals to their own components
