+++ src/components/Process/SuccessVoteModal.tsx
@@ -1,0 +1,72 @@
+import {
+  Box,
+  Button,
+  Link,
+  Modal,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  ModalOverlay,
+  Text,
+  useDisclosure,
+} from '@chakra-ui/react'
+import { environment } from '@vocdoni/chakra-components'
+import { useClient, useElection } from '@vocdoni/react-providers'
+import { useEffect, useState } from 'react'
+import { Trans, useTranslation } from 'react-i18next'
+import omniumLogo from '/assets/omnium.png'
+
+export const SuccessVoteModal = () => {
+  const { t } = useTranslation()
+  const { isOpen, onOpen, onClose } = useDisclosure()
+  const { votesLeft, election, voted } = useElection()
+  const { env } = useClient()
+
+  const [vLeft, setVLeft] = useState<number>(0)
+
+  useEffect(() => {
+    if (!vLeft && votesLeft >= 0) {
+      setVLeft(votesLeft)
+    }
+
+    if (vLeft && votesLeft < vLeft) {
+      setVLeft(votesLeft)
+      onOpen()
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [votesLeft, vLeft])
+
+  if (!election || !voted) return null
+
+  const verify = environment.verifyVote(env, voted)
+
+  return (
+    <Modal isOpen={isOpen} onClose={onClose}>
+      <ModalOverlay />
+      <ModalContent>
+        <ModalHeader>
+          <Text>{t('process.success_modal.title')}</Text>
+          <Box bgImage={omniumLogo} width='200px' mx='auto' />
+        </ModalHeader>
+        <ModalCloseButton />
+        <ModalBody>
+          <Trans
+            i18nKey='process.success_modal.text'
+            components={{
+              verify: <Link href={verify} target='_blank' />,
+              p: <Text mb={2} />,
+            }}
+          />
+        </ModalBody>
+
+        <ModalFooter mt={4}>
+          <Button onClick={onClose} variant='primary'>
+            {t('process.success_modal.btn')}
+          </Button>
+        </ModalFooter>
+      </ModalContent>
+    </Modal>
+  )
+}
